mm: return the address from page_mapped_in_vma()
ANBZ: #32353

commit 37bc2ff506b184411e4cc80f111c638b2b4c83d4 upstream.

The only user of this function calls page_address_in_vma() immediately
after page_mapped_in_vma() calculates it and uses it to return true/false.
Return the address instead, allowing memory-failure to skip the call to
page_address_in_vma().

Link: https://lkml.kernel.org/r/20240412193510.2356957-4-willy@infradead.org
Acked-by: Miaohe Lin <linmiaohe@huawei.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: Oscar Salvador <osalvador@suse.de>
Link: https://gitee.com/anolis/cloud-kernel/pulls/6687

+++ include/linux/rmap.h
@@ -723,7 +723,7 @@ enum rmp_flags {
 
 void remove_migration_ptes(struct folio *src, struct folio *dst, int flags);
 
-int page_mapped_in_vma(struct page *page, struct vm_area_struct *vma);
+unsigned long page_mapped_in_vma(struct page *page, struct vm_area_struct *vma);
 
 /*
  * rmap_walk_control: To control rmap traversing for specific needs
